api/requirements.txt
fastapi==0.104.1
Pillow==10.0.1
pydantic==2.5.2
pydantic_settings==2.1.0
torch==2.1.0
torchvision==0.16.0
uvicorn==0.24.0.post1
python-multipart==0.0.6

